--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -1,11 +1,11 @@
-module github.com/fixme_my_friend/hw10_program_optimization
+module github.com/TaoGunner/otus-hw/hw10_program_optimization
 
 go 1.22
 
-require github.com/stretchr/testify v1.7.0
+require github.com/stretchr/testify v1.9.0
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )