--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -3,13 +3,9 @@ package hw10programoptimization
 import (
 	"bufio"
 	"encoding/json"
-	"errors"
-	"fmt"
 	"io"
 	"strings"
 )
-
-var errTooMuchUsers = errors.New("too much users")
 
 // Остальные поля кроме Email в структуре User нас не интересуют (лишние аллокации).
 type User struct {
@@ -25,47 +21,18 @@ type User struct {
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
-
-type users [100_000]string
-
-// getUsers считывает построчно информацию о пользователях.
-func getUsers(r io.Reader) (result users, err error) {
-	usersCount := 0
-
-	// Считаем построчно в массив
+	result := make(DomainStat)
 	scanner := bufio.NewScanner(r)
-	for scanner.Scan() {
-		// Если пользователей больше 100_000 - ошибка
-		if usersCount >= len(result) {
-			return result, errTooMuchUsers
-		}
-
-		result[usersCount] = scanner.Text()
-		usersCount++
-	}
-
-	return
-}
-
-// countDomains подсчитывает количество доменов 1 уровня в электронной почте пользователей.
-func countDomains(u users, domain string) (result DomainStat, err error) {
-	result = make(DomainStat)
 	var user User
 
-	for idx := range u {
+	for scanner.Scan() {
 		// Пропускаем пользователей без совпадения домена
-		if !strings.Contains(u[idx], domain) {
+		if !strings.Contains(scanner.Text(), domain) {
 			continue
 		}
 
-		if err = json.Unmarshal([]byte(u[idx]), &user); err != nil {
-			return
+		if err := json.Unmarshal(scanner.Bytes(), &user); err != nil {
+			return result, err
 		}
 
 		if strings.HasSuffix(user.Email, domain) {